requirements.txt
GitPython==3.0.5
plotly==4.3.0
PyGithub==1.44.1

